@@ -2,58 +2,60 @@
   "name": "tcsn.io",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "author": {
     "name": "tcsn",
     "email": "hello@tcsn.io"
   },
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore",
+    "pages:dev": "wrangler pages dev dist --compatibility-date=2024-01-01",
+    "pages:build": "vite build",
+    "pages:deploy": "npm run pages:build && wrangler pages deploy dist --project-name=tcsnio",
+    "pages:deploy:prod": "npm run pages:build && wrangler pages deploy dist --project-name=tcsnio --branch=main",
+    "worker:dev": "wrangler dev workers/src/index.js",
+    "worker:deploy": "cd workers && wrangler deploy",
+    "worker:deploy:prod": "cd workers && wrangler deploy --env production",
+    "deploy:all": "npm run worker:deploy:prod && npm run pages:deploy:prod",
+    "d1:create": "wrangler d1 create tcsnio",
+    "d1:query": "wrangler d1 execute tcsnio --command",
+    "d1:migrations:create": "wrangler d1 migrations create tcsnio"
   },
   "dependencies": {
-    "@next/font": "13.1.2",
-    "@react-three/fiber": "^8.11.1",
-    "@reduxjs/toolkit": "^1.9.2",
-    "@types/next-auth": "^3.15.0",
-    "@types/node": "18.11.18",
-    "@types/react": "18.0.26",
-    "@types/react-dom": "18.0.10",
-    "@types/react-syntax-highlighter": "^15.5.6",
-    "@types/three": "^0.149.0",
-    "@vercel/analytics": "^0.1.8",
-    "axios": "^1.2.3",
-    "bcrypt": "^5.1.0",
-    "dotenv": "^16.0.3",
-    "eslint": "8.31.0",
-    "eslint-config-next": "13.1.2",
-    "googleapis": "^105.0.0",
+    "@fortawesome/fontawesome-svg-core": "^6.5.0",
+    "@fortawesome/free-brands-svg-icons": "^6.5.0",
+    "@fortawesome/free-regular-svg-icons": "^6.5.0",
+    "@fortawesome/free-solid-svg-icons": "^6.5.0",
+    "@fortawesome/vue-fontawesome": "^3.0.5",
+    "animejs": "^4.2.2",
+    "axios": "^1.6.0",
+    "dotenv": "^16.6.1",
+    "google-auth-library": "^10.3.0",
+    "js-cookie": "^3.0.5",
+    "jsonwebtoken": "^9.0.2",
+    "leaflet": "^1.9.4",
+    "lucide-vue-next": "^0.294.0",
     "moment": "^2.29.4",
-    "mongoose": "^6.8.4",
-    "next": "13.1.2",
-    "next-auth": "^4.18.8",
-    "nodemailer": "^6.9.0",
-    "react": "18.2.0",
-    "react-code-blocks": "0.0.9-0",
-    "react-dom": "^18.2.0",
-    "react-feather": "^2.0.10",
-    "react-icons": "^4.7.1",
-    "react-redux": "^8.0.5",
-    "react-syntax-highlighter": "^15.5.0",
-    "three": "^0.149.0",
-    "typescript": "4.9.4"
+    "openai": "^5.20.0",
+    "pinia": "^2.1.0",
+    "prismjs": "^1.29.0",
+    "three": "^0.158.0",
+    "vue": "^3.4.0",
+    "vue-router": "^4.2.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.12",
-    "autoprefixer": "^10.4.13",
-    "babel-loader": "^8.3.0",
-    "mongodb": "^6.10.0",
-    "postcss": "^8.4.21",
-    "react-openlayers": "^0.3.0",
-    "tailwindcss": "^3.2.6"
-  },
-  "resolutions": {
-    "next": "13.1.2"
+    "@cloudflare/workers-types": "^4.20231218.0",
+    "@vitejs/plugin-vue": "^5.0.0",
+    "autoprefixer": "^10.4.16",
+    "eslint": "^8.52.0",
+    "eslint-plugin-vue": "^9.17.0",
+    "postcss": "^8.4.31",
+    "sass": "^1.69.0",
+    "tailwindcss": "^3.3.0",
+    "vite": "^5.4.0",
+    "wrangler": "^4.42.1"
   }
 }
